ML/_nv/yolov1/requirements.txt

pytorch
pandas
pillow